+++ frontend/src/components/dating/DatingOptionMultipleChoose.tsx
@@ -1,9 +1,8 @@
 /* A custom multiple choose button component. */
 import { chakra, useCheckbox, Flex, Box, Text } from "@chakra-ui/react"
-import { useState } from "react"
 
-export function DatingOptionMultiChose(props: any) {
-    const { state, getCheckboxProps, getInputProps, getLabelProps, htmlProps } = useCheckbox(props)
+export function DatingOptionMultipleChoose(props: any) {
+    const { state, getCheckboxProps, getLabelProps, htmlProps } = useCheckbox(props)
 
     // For multiple chose of faculties
     return (
@@ -25,10 +24,8 @@ export function DatingOptionMultiChose(props: any) {
             onClick={(e: any) => {
                 props.handelClick(props.value)
                 console.log("Original : " + props.value)
-                // e.preventDefault()
             }}
         >
-            {/* <input {...getInputProps()} hidden /> */}
             <Flex alignItems="center" justifyContent="center" border="2px solid" borderColor="orange.500" w={4} h={4} {...getCheckboxProps()}>
                 {state.isChecked && <Box w={2} h={2} bg="orange.500" />}
             </Flex>
