+++ frontend/src/pages/dating/option.tsx
@@ -1,30 +1,9 @@
-import {
-    Heading,
-    Text,
-    Box,
-    RangeSlider,
-    RangeSliderTrack,
-    RangeSliderFilledTrack,
-    RangeSliderThumb,
-    Stack,
-    Center,
-    Checkbox,
-    Button,
-    RangeSliderMark,
-    SimpleGrid,
-    AccordionItem,
-    AccordionButton,
-    AccordionPanel,
-    AccordionIcon,
-    Accordion,
-    Tooltip,
-} from "@chakra-ui/react"
-import React, { useEffect, useState } from "react"
+import { Heading, Text, Box, Stack, Center, Button, SimpleGrid, useRadioGroup, useCheckboxGroup } from "@chakra-ui/react"
+import { useEffect, useState } from "react"
 import { DatingOptionRadioBox } from "../../components/dating/DatingOptionRadioBox"
-import { useRadioGroup } from "@chakra-ui/react"
 import DatingAppBody from "../../components/dating/DatingAppBody"
-import { useCheckboxGroup } from "@chakra-ui/react"
-import { DatingOptionMultiChose } from "../../components/dating/DatingOptionMultiChose"
+import DatingOptionRangeSlider from "../../components/dating/DatingOptionRangeSlider"
+import DatingOptionAccordion from "../../components/dating/DatingOptionAccordion"
 
 declare global {
     var age: number[], gender: string, faculty: string[], useAge: boolean
@@ -74,58 +53,10 @@ const DatingOption = () => {
     useEffect(() => {
         setSelectedFac(faculties)
     }, [])
-    function handleAge() {
-        //Passing data + debug
-        globalThis.age = sliderValue
-        // console.log(globalThis.age)
-    }
-
-    function handleCheck() {
-        //Passing data + debug
-        globalThis.useAge = useAgeValue
-        // console.log(globalThis.useAge)
-    }
 
     function handleGender(gender: string) {
         //Passing data
         setSelected(gender)
-    }
-
-    function handleFac(fac: any) {
-        let arr: string[] = selectedFac
-        if (fac === "All Faculty") {
-            if (arr.includes(fac)) {
-                setSelectedFac([])
-            } else setSelectedFac(faculties)
-            return
-        }
-        console.log("This arr: " + arr)
-        if (!arr.includes(fac)) {
-            arr = [...arr, fac]
-            arr.sort()
-            setSelectedFac([...arr])
-            console.log("array: " + selectedFac)
-            console.log("This add? :" + arr.indexOf(fac))
-        } else {
-            // filter?
-            arr = arr.filter((item) => item !== fac)
-            setSelectedFac([...arr])
-
-            console.log("This remove? :" + arr.splice(arr.indexOf(fac), arr.indexOf(fac) + 1))
-        }
-        let arrWithoutAllfact = faculties.filter((item) => item !== faculties[0])
-        let isAll = true
-        arrWithoutAllfact.forEach((item) => {
-            if (!arr.includes(item)) {
-                isAll = false
-            }
-        })
-        if (isAll) {
-            setSelectedFac([faculties[0], ...arr])
-        } else {
-            setSelectedFac(arr.filter((item) => item !== faculties[0]))
-        }
-        console.log("This :" + arr)
     }
 
     function handleSubmit() {
@@ -135,9 +66,6 @@ const DatingOption = () => {
         globalThis.gender = selected
         globalThis.faculty = selectedFac
         console.log(selectedFac)
-        if (globalThis.faculty[0] == "All Faculty") {
-            console.log("All Fac")
-        }
         alert(
             "Age min =" +
                 globalThis.age[0] +
@@ -169,53 +97,12 @@ const DatingOption = () => {
                                 Age Preference
                             </Text>
                         </Box>
-                        <Center bg="orange.200" h={20} borderRadius="20px">
-                            {/* Is user use age as criteria? */}
-                            <Tooltip label="Use age as a criteria?" aria-label="A tooltip">
-                                <span tabIndex={0}>
-                                    <Checkbox
-                                        colorScheme="white"
-                                        defaultChecked={globalThis.useAge}
-                                        p="30px"
-                                        size="lg"
-                                        onChange={() => {
-                                            handleCheck()
-                                            setUseAgeValue(!useAgeValue)
-                                        }}
-                                    ></Checkbox>
-                                </span>
-                            </Tooltip>
-
-                            {/* Age min and Age max */}
-                            <RangeSlider
-                                aria-label={["min", "max"]}
-                                min={18}
-                                max={40}
-                                defaultValue={[19, 25]}
-                                onChange={(val) => {
-                                    handleAge()
-                                    setSliderValue(val)
-                                }}
-                                ml={"20px"}
-                                mr={"45px"}
-                            >
-                                <RangeSliderMark value={sliderValue[0]} textAlign="center" color="white" mt="-10" ml="-5" w="12">
-                                    {sliderValue[0]}
-                                </RangeSliderMark>
-                                <RangeSliderMark value={sliderValue[1]} textAlign="center" color="white" mt="-10" ml="-5" w="12">
-                                    {sliderValue[1]}
-                                </RangeSliderMark>
-                                <RangeSliderTrack bg="red.100">
-                                    <RangeSliderFilledTrack bg="tomato" />
-                                </RangeSliderTrack>
-                                <RangeSliderThumb boxSize={6} index={0}>
-                                    <Box color="tomato" />
-                                </RangeSliderThumb>
-                                <RangeSliderThumb boxSize={6} index={1}>
-                                    <Box color="tomato" />
-                                </RangeSliderThumb>
-                            </RangeSlider>
-                        </Center>
+                        <DatingOptionRangeSlider
+                            sliderValue={sliderValue}
+                            useAgeValue={useAgeValue}
+                            setUseAgeValue={setUseAgeValue}
+                            setSliderValue={setSliderValue}
+                        />
                     </Box>
                     <Box>
                         <Box pb={5}>
@@ -243,127 +130,12 @@ const DatingOption = () => {
                             </Text>
                         </Box>
                         <Box>
-                            <Accordion allowToggle flex="left">
-                                <AccordionItem>
-                                    <h2>
-                                        <AccordionButton
-                                            bg={"orange.200"}
-                                            color="white"
-                                            borderRadius="20px"
-                                            _expanded={{ bg: "orange.200", color: "white" }}
-                                            _hover={{ bg: "orange.300", color: "white", border: "orange.800" }}
-                                        >
-                                            <Box textAlign="left" borderRadius="full">
-                                                Selected Faculty
-                                            </Box>
-                                            <AccordionIcon />
-                                        </AccordionButton>
-                                    </h2>
-                                    <AccordionPanel pb={4}>
-                                        <Stack>
-                                            {/* <Text>You have select from: {selectedFac.sort().join(" , ")}</Text> */}
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[0] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[0])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[1] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[1])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[2] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[2])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[3] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[3])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[4] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[4])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[5] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[5])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[6] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[6])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[7] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[7])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[8] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[8])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[9] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[9])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[10] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[10])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[11] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[11])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[12] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[12])}
-                                            />
-                                            <DatingOptionMultiChose
-                                                {...getCheckboxProps({ value: faculties[13] })}
-                                                handelClick={(e: any) => {
-                                                    handleFac(e)
-                                                }}
-                                                isChecked={selectedFac.includes(faculties[13])}
-                                            />
-                                        </Stack>
-                                    </AccordionPanel>
-                                </AccordionItem>
-                            </Accordion>
+                            <DatingOptionAccordion
+                                faculties={faculties}
+                                selectedFac={selectedFac}
+                                setSelectedFac={setSelectedFac}
+                                getCheckboxProps={getCheckboxProps}
+                            />
                         </Box>
                     </Box>
                 </SimpleGrid>
