+++ frontend/src/pages/rating.tsx
@@ -1,5 +1,3 @@
-import { Box } from "@chakra-ui/react"
-import React from "react"
 import AppBody from "../components/share/app/AppBody"
 
 const Rating = () => {
